// src/app/component/types/BinMaster.tsx
export interface BinMaster {
    binID: number;
    whsCode: string;
    binLocCode: string;
    sL1Code: string;
    sL2Code: string;
    sL3Code: string;
    sL4Code: string;
    sL5Code: string;
    height: number;
    width: number;
    length: number;
    filter1: string;
    filter2: string;
    filter3: string;
    quantity: number;
    level: number;
    active: boolean;
    userSign: string;
}